Uploading the mwb file from MySQL Workbench
Created the foundation of my application. Starting with a One-to-Many relationship so that one user can have multiple children, but for now a child can only belong to one user. I plan to open it up to other users where multiple users can have access to edit/add to a child's journal, but that will come at a later date.